Ported changes from ZDoom revision 1735-1747 and from GZDoom revision 406-409: - Added A_Mushroom compatibility option for Dehacked. - Added Gez's submission for interhubamount DECORATE property. - Fixed: The big endian version of PalEntry did not add the DWORD d field. - Fixed: TObjPtr did not use a union to map its 2 pointers together. - Added a compatibility mode for A_Mushroom. For DECORATE it is an additional parameter but to force it for Dehacked mods some minor hacks using the Misc1 variable were needed. - Moved the terget->velz assignment to the end of A_VileAttack to remove the influence of vertical thrust from the radius attack, since ZDoom does explosions in three dimensions, but Doom only did it in two. - Fixed: The last three parameters to A_VileAttack had their references off by one. Most notably as a result, the blast radius was used as the thrust, so it sent you flying far faster than it should have. - Restored the reactiontime countdown for A_SpawnFly, since some Dehacked mod could conceivable rely on it. The deadline is now kept in special2 and used in preference as long as it is non-zero. - Removed -fno-strict-aliasing from the GCC flags for ZDoom and fixed the issues that caused its inclusion. Is an optimized GCC build any faster for being able to use strict aliasing rules? I dunno. It's still slower than a VC++ build.
  I did run into two cases where TAutoSegIterator caused intractable problems with breaking strict aliasing rules, so I removed the templating from it, and the caller is now responsible for casting the probe value from void *.

- Removed #include "autosegs.h" from several files that did not need it (in particular, dobject.h when not compiling with VC++).
- Changed A_SpawnFly to do nothing if reactiontime is 0. Reactiontime was originally a counter, so if it started at 0, A_SpawnFly would effectively never spawn anything. Fixes Dehacked patches that use A_SpawnSound to play a sound without shooting boss cubes, since it also calls A_SpawnFly.
- Joystick devices now send key up events for any buttons that are held down when they are destroyed.
- Changed the joystick enable-y menu items to be nonrepeatable so that you can't create several device scans per second. Changing them with a controller is also disabled so that you can't, for example, toggle XInput support using an XInput controller and have things go haywire when the game receives an infinite number of key down events when the controller is reinitialized with the other input system.
- Changed menu input to use a consolidated set of buttons, so most menus can be navigated with a controller as well as a keyboard.
- Changed the way that X/Y analog axes are converted to digital axes. Previously, this was done by checking if each axis was outside its deadzone. Now they are checked together based on their angle, so straight up/down/ left/right are much easier to achieve.
- Changed: DEHACKED parsing is disabled now when a user supplied DEHSUPP lump is found. This mimics the old behavior which also disabled DEHACKED when the DEHSUPP lump was incompatible with the current engine. This behavior is needed to ensure that WADs that contain a ZDaemon-exclusive DEHSUPP lump continue to work as intended.
- fixed namespace assignment for gl_vertices array.

+++ src/gl/common/glc_sections.cpp
@@ -555,7 +555,7 @@ void DumpSection(int no, FGLSection *sect)
 				double d2 = sqrt(dx2*dx2+dy2*dy2);
 
 				Printf(PRINT_LOG, "\t\tLinedef %d, %s: Start (%1.2f, %1.2f), End (%1.2f, %1.2f)", 
-					ln->linedef - lines, (ln->sidedef - sides) == ln->linedef->sidenum[0]? "front":"back",
+					ln->linedef - lines, (unsigned int)(ln->sidedef - sides) == ln->linedef->sidenum[0]? "front":"back",
 					ln->start->x/65536.f, ln->start->y/65536.f,
 					ln->end->x/65536.f, ln->end->y/65536.f);
